User Interface changes
in the client.py file, the main changes have been in the user interface handler and the user interface. There is now a login window. If you wish to login using this you will have to update some code.

For starters you will have to change the I.P. address to the address of the PC that your server is running on. You can also change the port to something you like to use better if you want, although its not needed.

you will not be able to use the functions for sending emails as the email bots password is not shown. Once the finished version is complete all that information will be available.

when you start up the program you will also notice that there are no books yet. that is because I have not added the .dat files.

You will have to add your own books to the program and make sure that it saves to your computer.